requirements.txt
django==3.1
pillow==8.0
psycopg2-binary==2.8.5
djangorestframework==3.11.0
python-decouple==3.4
django-phonenumber-field==5.0.0
phonenumbers==8.12.19
djangorestframework-simplejwt
twilio
django-modeltranslation